BJN-23: Refactored socket code into approciate files and fixed port mapping for frontend
