+++ docker-compose.yml
@@ -56,7 +56,7 @@ services:
         - VITE_FIREBASE_AUTH_EMULATOR_URL=${VITE_FIREBASE_AUTH_EMULATOR_URL}
         - VITE_SERVER_URL=${VITE_SERVER_URL}
     ports:
-      - '3000:3000'
+      - '8080:8080'
     depends_on:
       - backend
       - firebase
